% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_my.R
\name{days_my}
\alias{days_my}
\title{days_my}
\usage{
days_my(data, grid = FALSE, density = FALSE, line = FALSE)
}
\description{
A function for analysis the milk yield of the herd.
}
\examples{
days_my(data = dairyCattle::read_cattle("cattle_data.xls", drop.zero = TRUE, add = TRUE), grid = FALSE, density = FALSE, line = FALSE)
}
\keyword{cattle}
\keyword{dairy}
\keyword{milk}
\keyword{yield}
