% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_my.R
\name{dim_my}
\alias{dim_my}
\title{dim_my}
\usage{
dim_my(data, grid = FALSE, density = FALSE, line = FALSE, text = FALSE)
}
\description{
A function for analysis the milk yield of the herd.
}
\examples{
dim_my(data = dairyCattle::read_cattle("cattle_data.xls", drop.zero = TRUE, add = TRUE), grid = FALSE, density = FALSE, line = FALSE, text = FALSE)
}
\keyword{cattle}
\keyword{dairy}
\keyword{days-in-milk}
\keyword{milk}
\keyword{yield}
