% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat_prot.R
\name{fat_prot}
\alias{fat_prot}
\title{fat_prot}
\usage{
fat_prot(data, grid = FALSE, density = FALSE, line = FALSE,
  text = FALSE)
}
\description{
A function for the milk fat-protein analysis of the herd.
}
\examples{
fat_prot(data = dairyCattle::read_cattle("cattle_data.xls", drop.zero = TRUE, add = TRUE), grid = FALSE, density = FALSE, line = FALSE, text = FALSE)
}
\keyword{cattle}
\keyword{dairy}
\keyword{fat}
\keyword{milk}
\keyword{protein}
