% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_cattle.R
\name{gs_cattle}
\alias{gs_cattle}
\title{gs_cattle}
\usage{
gs_cattle(path, drop.zero = FALSE, add = FALSE)
}
\arguments{
\item{path}{google spread sheet file.}

\item{drop.zero}{remove the dataset that milk yeild == 0.}

\item{add}{add some columns for additional analysis.}
}
\description{
read_cattle for google spread sheet file
}
\examples{
df <- googlesheets::gs_title("example") \%>\% googlesheets::gs_read()
gs_cattle(path = df, drop.zero = TRUE)
}
\keyword{cattle}
\keyword{dairy}
