% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_mp_tbl.R
\name{mun_mp_tbl}
\alias{mun_mp_tbl}
\title{mun_mp_tbl}
\usage{
mun_mp_tbl(data, dataframe = FALSE)
}
\description{
A function makes list for the milk mun-protein analysis
}
\examples{
mun_mp_tbl(data = dairyCattle::read_cattle("cattle_data.xls", drop.zero = TRUE, add = TRUE))
}
\keyword{MUN}
\keyword{cattle}
\keyword{dairy}
