% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parturition.R
\name{parturition}
\alias{parturition}
\title{parturition}
\usage{
parturition(data)
}
\description{
A time table for parturition of herd.
}
\examples{
parturition(data = dairyCattle::read_cattle("cattle_data.xls", drop.zero = TRUE, add = TRUE))
}
\keyword{dairy}
\keyword{parturition}
