% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parturition_tbl.R
\name{parturition_tbl}
\alias{parturition_tbl}
\title{parturition_tbl}
\usage{
parturition_tbl(data, year = NULL, month = NULL)
}
\description{
A time table for parturition of herd.
}
\examples{
df <- dairyCattle::read_cattle("cattle_data.xls", drop.zero = FALSE, add = TRUE)
parturition_tbl(data = df)
parturition_tbl(data = df, year = 2018, month = 8)
}
\keyword{dairy}
\keyword{parturition}
