% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cattle.R
\name{read_cattle}
\alias{read_cattle}
\title{read_cattle}
\usage{
read_cattle(file, na.rm = FALSE)
}
\arguments{
\item{file}{file path}

\item{na.rm}{remove the dataset that milk yeild == 0.}
}
\description{
This import the cattle data from http://www.aiak.or.kr.
}
\examples{
read_cattle("test.xlsx",na.rm=TRUE)
read_cattle("test.xlsx",na.rm=FALSE)
}
\keyword{cattle}
\keyword{dairy}
