% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mun_mp.R
\name{mun_mp}
\alias{mun_mp}
\title{mun_mp}
\usage{
mun_mp(data)
}
\description{
A function for mun-milk protein analysis
}
\examples{
mun_mp(data = dairyCattle::read_cattle("cattle_data.xls", drop.zero = TRUE, add = TRUE))
}
\keyword{MUN}
\keyword{cattle}
\keyword{dairy}
