% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cattle.R
\name{read_cattle}
\alias{read_cattle}
\title{read_cattle}
\usage{
read_cattle(path, drop.zero = FALSE, add = FALSE)
}
\arguments{
\item{path}{path to the xls/xlsx/csv/txt file.}

\item{drop.zero}{remove the dataset that milk yeild == 0.}

\item{add}{add some columns for additional analysis.}
}
\description{
This import the cattle data from http://www.aiak.or.kr.
}
\examples{
read_cattle(path="result.xls",drop.zero=TRUE)
read_cattle(path="result.xls",drop.zero=FALSE,add=TRUE)
}
\keyword{cattle}
\keyword{dairy}
